% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_botometer.R
\name{install_botometer}
\alias{install_botometer}
\title{Install botometer}
\usage{
install_botometer(method = "auto", conda = "auto")
}
\arguments{
\item{method}{A string, defaults to \code{"auto"}}

\item{conda}{A string, defaults to \code{"auto"}}
}
\description{
Install the Python module botometer
}
\details{
This function is a wrapper for calling \code{py_install()} from 
the \code{reticulate} pacakge to install the Python module 
\code{botometer}.
}
\references{
The source of this code is 
\url{https://github.com/rstudio/reticulate/blob/master/vignettes/package.Rmd}.
}
\author{
Ryan T. Moore
}
