% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_botscan.R
\name{setup_botscan}
\alias{setup_botscan}
\title{setup_botscan}
\usage{
setup_botscan(
  rapidapi_key,
  consumer_key,
  consumer_secret,
  access_token,
  access_token_secret
)
}
\arguments{
\item{rapidapi_key}{A string representing the RapidAPI key}

\item{consumer_key}{A string representing the consumer key}

\item{consumer_secret}{A string representing the consumer secret}

\item{access_token}{A string representing the access token}

\item{access_token_secret}{A string representing the access token secret}
}
\value{
A list of length two with components
\item{rapidapi_key}{The RapidAPI key}
\item{twitter_app_auth}{A list of length four, including the consumer_key, 
consumer_secret, access_token, and access_token_secret}
}
\description{
Initialize Keys, Tokens, and Secrets for Twitter Access
}
\details{
Takes a Twitter query and produces the percentage of users within that
conversation that are likely to be bots.
}
\examples{
\dontrun{bom <- setup_botscan("a", "b", "c", "d", "e")}

# The above example fails, since each key, secret, or token must be created
# by the user via ...

}
\author{
Ryan T. Moore
}
