% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botscan.R
\name{botscan}
\alias{botscan}
\title{botscan}
\usage{
botscan(x)
}
\arguments{
\item{x}{A Twitter search query in quotation marks.}
}
\value{
Percentage of users within the requested conversation that are
estimated by botornot to have at least a 90% chance of being a bot.

Examples fail unless you have created and installed Twitter tokens, per
instructions provided at http://rtweet.info/articles/auth.html.
}
\description{
Scans Twitter Conversations For Bots
}
\details{
Takes a Twitter query and produces the percentage of users within that
conversation that are likely to be bots.
}
\examples{
\dontrun{botscan("#rtweets")}
\dontrun{botscan("trump")}
}
\author{
Kurt Wirth
}
