% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botscan.R
\name{botscan}
\alias{botscan}
\title{botscan}
\usage{
botscan(x, threshold = 0.899, user_level = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{A string representing a Twitter search query, in quotation marks.}

\item{threshold}{A number between zero and one that determines which botornot 
probability threshold to return. Default is set at 0.899. Only users estimated to be 
more likely than the threshold provided will be regarded as a bot.}

\item{user_level}{A logical that determines whether to analyze
conversation-level or user-level data. Default is set to \code{FALSE}, understood
as analyzing conversation-level data.}

\item{verbose}{A logical that determines whether to print periodic progress
updates.}
}
\value{
Proportion of users within the requested conversation that are 
estimated by botometer to be bots.

When \code{user_level} is set to \code{TRUE}, the value returned is the percentage of
users within the conversation that are estimated to be bots. When
\code{user_level} is set to \code{FALSE}, the value returned is the percentage of the
conversation that is estimated to be authored by bots.
}
\description{
Scans Twitter Conversations For Bots
}
\details{
Takes a Twitter query and produces the proportion of users within that
conversation that are likely to be bots.
}
\examples{
\dontrun{botscan("#rtweets")}
\dontrun{botscan("trump")}

##The above examples fail unless you have created and installed Twitter 
##tokens, per instructions provided at http://rtweet.info/articles/auth.html.

}
\author{
Kurt Wirth and Ryan T. Moore
}
