% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botscan.R
\name{botscan}
\alias{botscan}
\title{botscan}
\usage{
botscan(x)
}
\arguments{
\item{x}{A Twitter search query in quotation marks.}
}
\value{
A number
}
\description{
Scans Twitter Conversations For Bots
}
\details{
Takes a Twitter query and produces the percentage of users within that
conversation that are likely to be bots.
}
\examples{
\dontrun{botscan("#rtweets")}
\dontrun{botscan("trump")}
}
\author{
Kurt Wirth
}
