% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botscan.R
\name{botscan}
\alias{botscan}
\title{botscan}
\usage{
botscan(x, timeout = 30, threshold = 0.43, stream = TRUE,
  n_tweets = 1000, retweets = FALSE, parse = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{A string representing a Twitter search query, in quotation marks.}

\item{timeout}{A number representing the number of seconds the user wishes to
stream tweets on the given search term. This is only applied when using the
default STREAM Twitter API. Default is 30 seconds.}

\item{threshold}{A number between zero and one that determines which botornot 
probability threshold to return. Default is set at 0.43. Only users estimated 
to be more likely than the threshold provided will be regarded as a bot.}

\item{stream}{A logical indicating whether the streaming API or the search API 
is queried.  Defaults to \code{TRUE}, using the streaming API.}

\item{n_tweets}{A number representing how many tweets to extract.}

\item{retweets}{A logical specifying whether to include retweets in 
the set of tweets to be extracted.  Defaults to \code{FALSE}.}

\item{parse}{A logical specifying whether to automatically parse data into
structured data when querying the streaming API. Defaults to \code{TRUE}, 
making results more usable for most users. Setting to \code{FALSE} will
produce less structured data but allow for faster processing for very large
data projects.}

\item{verbose}{A logical that determines whether to print periodic progress
updates.}
}
\value{
A list of length 3: \enumerate{
   \item Dataframe including all data. This includes original data for each tweet as
         well as BotOMeter scores for each.
   \item User-level estimate for the given search.
   \item Conversation-level estimate for the given search.
}

When \code{user_level} is set to \code{TRUE}, the value returned is the percentage of
users within the conversation that are estimated to be bots. When
\code{user_level} is set to \code{FALSE}, the value returned is the percentage of the
conversation that is estimated to be authored by bots.
}
\description{
Scans Twitter Conversations For Bots
}
\details{
Takes a Twitter query and produces the proportion of users within that
conversation that are likely to be bots.

To get a sample of tweets from the stream API, set \code{x = ""}.
}
\examples{
\dontrun{botscan("#rtweets")}
\dontrun{botscan("trump")}

##The above examples fail unless you have created and installed Twitter 
##tokens, per instructions provided at http://rtweet.info/articles/auth.html.

}
\author{
Kurt Wirth and Ryan T. Moore
}
